+++ CHANGELOG.rst
@@ -14,6 +14,9 @@ functions, via the ``out`` argument.
 
 Full list of changes:
 
+* **Backwards incompatible** ``PageObjectRegistry`` is no longer supporting
+  dict-like access.
+
 * Official support for Python 3.11.
 
 * New ``@web_poet.field(out=[...])`` argument which allows to set output
@@ -42,10 +45,11 @@ Full list of changes:
 
 * ``PageObjectRegistry`` is replaced with ``RulesRegistry``; its API is changed:
 
-    * dict-like API is deprecated, and can be dropped in the future;
+    * **backwards incompatible** dict-like API is removed;
     * ``search_overrides`` method is renamed to ``search``;
     * ``get_overrides`` method is renamed to ``get_rules``;
-    * ``from_override_rules`` method is renamed to ``from_apply_rules``.
+    * ``from_override_rules`` method is deprecated;
+      use ``RulesRegistry(rules=...)`` instead.
 
 * Typing improvements.
 * Documentation, test, and warning message improvements.
@@ -58,7 +62,7 @@ Deprecations:
 * The ``OverrideRule`` class is now deprecated. Use ``ApplyRule`` instead.
 * ``PageObjectRegistry`` is now deprecated. Use ``RulesRegistry`` instead.
 * The ``from_override_rules`` method of ``PageObjectRegistry`` is now deprecated.
-  Use ``from_apply_rules`` instead.
+  Use ``RulesRegistry(rules=...)`` instead.
 * The ``PageObjectRegistry.get_overrides`` method is deprecated.
   Use ``PageObjectRegistry.get_rules`` instead.
 * The ``PageObjectRegistry.search_overrides`` method is deprecated.